Move Function Attachment Into Application
